requirements.txt
discord.py
python-dotenv
asyncio
openai